[Close #177] Extracted destroy-session action.
New unbuilt Gem integrated into app, replacing previously-existing
destroy-session action.

RSpec: 572+2+5+39+24+34+30 (706) examples, 0 failures
RuboCop: 198+4+4+12+5+8+5 (236) files inspected, no offenses detected
