+++ lib/newpoc-action-session-destroy/newpoc-action-session-destroy.gemspec
@@ -18,6 +18,9 @@ Gem::Specification.new do |spec|
   spec.test_files    = spec.files.grep(%r{^(test|spec|features)/})
   spec.require_paths = ["lib"]
 
+  spec.add_dependency 'wisper', '~> 1.6'
+
+  spec.add_development_dependency 'wisper_subscription', '>= 0.2.0'
   spec.add_development_dependency "bundler", "~> 1.7"
   spec.add_development_dependency "rake", "~> 10.0"
   spec.add_development_dependency "rspec"
